requirements.txt
sentence-transformers>=2.2.0
plotly>=5.14.0
umap-learn>=0.5.3
scikit-learn>=1.2.0
numpy>=1.24.0
dash>=3.0.0
dash-bootstrap-components>=2.0.0
flask>=2.3.0
flask-cors>=4.0.0
